sdpyqt0.0.0.2-bt-hotfix
There was a bug in the downloads manager that I attempted to fix, and it seems to be resolved now.

Previously, when you cleared the download history, if there were no entries in the history, the browser would crash. Moreover, trying to restart the browser would cause it to keep crashing.

Now, the app crashes only on the second attempt, but after that, it no longer crashes when you restart it.